Recursively search local sources
Sources can now be loaded recursively assuming the code implementing the source
supports it. Currently, the local sources are the only things supported and http
sources are very broken, but that can be fixed easily.

Local handles sources recursively and constantly prepends namespaces. Note that
this must be implemented by the source (local, http, etc.) and, as a result,
could be skipped by the right source.