Point to other assets using assets
Every non-playlist asset was previously treated as a video and nothing else. It
was impossible to link one asset to another. This fixes that problem. You can do
something like "main: A:main" if you want main to point to main in the source A.
These links can go as deep as needed.